feat: async all node next cycle calculation
When we calculate the next cycle in Node class, it change right away and update node.IsAlive. As a result  later node calculation is wrong, as they are not using what the last 'cell' status.
Solution here is we save the next cycle status is Grid._board, and apply them after all node is calculated.